+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Unreleased
 ### Additions:
 
 - Add `StructureObject::as_repairable` matching related functions for other object traits
+- Add implementations of `TryFrom<AccountPowerCreep>` and `AsRef<AccountPowerCreep>` for
+  `PowerCreep`
+- Add implementation of `From<PowerCreep>` for `AccountPowerCreep`
 
 0.20.0 (2024-01-08)
 ===================