+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 - Add implementations of `TryFrom<AccountPowerCreep>` and `AsRef<AccountPowerCreep>` for
   `PowerCreep`
 - Add implementation of `From<PowerCreep>` for `AccountPowerCreep`
+- Add `constants::extra::NOTIFY_MAX_LENGTH` with the length of allowed input to `game::notify`
 
 0.20.0 (2024-01-08)
 ===================